requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# OpenAI
openai==1.10.0

# PDF Generation
reportlab==4.0.9

# Monitoring
prometheus-fastapi-instrumentator==6.1.0
sentry-sdk[fastapi]==1.39.2

# Production
gunicorn==21.2.0
uvicorn[standard]==0.27.0

pytest>=7.4.3
pytest-asyncio==0.21.1
httpx>=0.25.2
secure>=0.3.0
prometheus-client>=0.19.0 